Implemented part of the Whatsapp api
