+++ api/utils/.env
@@ -6,7 +6,7 @@ SERVER_BASE_URL=http://localhost/kiosk/
 DB_NAME=kiosk
 PAYBOX_BEARER_TOKEN=f6fea1ab-76bc-4390-9a46-a6de4b0c2de8
 FB_GRAPH_VERSION=v15.0
-REDIRECT_URL=http://localhost/Kiosk/api/whatsapp_api/redirect_url.php
+REDIRECT_URL=http://localhost/Kiosk/api/api/redirect_url.php
 FB_APP_STATE=fb_state_check
 FB_GRAPH_DOMAIN= https://graph.facebook.com/
 FB_MAIN_DOMAIN=https://www.facebook.com/
