--- api/whatsapp_api/processors/whatsapp_api_processor.php
@@ -1,10 +1,0 @@
-<?php
-require("../controllers/whatsapp_api.php");
-require("../../whatsapp_api/redirect_url.php");
-
-$app_Id = $_GET["app_id"];
-$app_Secret = $_GET["app_secret"];
-
-
-
-print_r(facebookLogin($app_Id, $app_Secret));
